Move tests to proper classes, polishing tests
