+++ src/test/java/redis/clients/jedis/tests/commands/HashesCommandsTest.java
@@ -417,4 +417,29 @@ public class HashesCommandsTest extends JedisCommandTestBase {
     assertFalse(bResult.getResult().isEmpty());
   }
 
+  @Test
+  public void testHstrLen_EmptyHash() {
+    Long response = jedis.hstrlen("myhash", "k1");
+    assertEquals(0l, response.longValue());
+  }
+
+  @Test
+  public void testHstrLen() {
+    Map<String, String> values = new HashMap<String, String>();
+    values.put("key", "value");
+    jedis.hmset("myhash", values);
+    Long response = jedis.hstrlen("myhash", "key");
+    assertEquals(5l, response.longValue());
+
+  }
+
+  @Test
+  public void testBinaryHstrLen() {
+    Map<byte[], byte[]> values = new HashMap<byte[], byte[]>();
+    values.put(bbar, bcar);
+    jedis.hmset(bfoo, values);
+    Long response = jedis.hstrlen(bfoo, bbar);
+    assertEquals(4l, response.longValue());
+  }
+
 }
